Debug input field clearing
Input fields were clearing because thier initial values were all
set to ''. I solved this by declaring the values in the App.js
component state instead, and then pass them into MyForm.js as a
prop and call them in the initial values object as
`age: formData.age`.

All functionallity is currently working in my app as intended.
Data is still sent to the resume page without error, the form
populates with the values in state when re-rendered so the user
can edit them, and is correctly passed into parent component
state and resume component when resubmitted.