Add work experience fields and validation
All fields, validation, and state management working as expected.
Unfortunately working with nested objects in useFormik was not
possible. In future projects, <Formik /> is likely the wiser choice,
as there is better support for it.

Because of this, more hard-coding occured than desired but the form
component itself is still pretty modular and could be re-purposed
to future projects with little effort.